+++ NRPFarmod/ContentManager/TextureMananger.cs
@@ -1,15 +1,36 @@
 ﻿using MelonLoader;
+using NRPFarmod.UIHelper;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
+using UnityEngine.Experimental.UIElements;
 
 namespace NRPFarmod.ContentManager {
+
     public static class TextureMananger {
 
+        #region AboutUI
 
+
+
+
+        #region AboutUI Position
+
+
+
+        #endregion
+        #endregion
+
+        #region Create Texture
+        /// <summary>
+        /// Lädt die Texture
+        /// </summary>
+        /// <param name="tmp"></param>
+        /// <param name="data"></param>
         public static void CreateTexture(ref Texture2D? tmp, byte[] data) {
 
             try {
@@ -22,7 +43,7 @@ namespace NRPFarmod.ContentManager {
                 uint width = enumerator.Current; enumerator.MoveNext();
                 uint height = enumerator.Current; enumerator.MoveNext();
 
-                MelonLogger.Msg($"Texture: \u001b[32m{width}\u001b[0mx\u001b[32m{height}\u001b[0m");
+                //MelonLogger.Msg($"Texture: \u001b[32m{width}\u001b[0mx\u001b[32m{height}\u001b[0m");
 
                 tmp = new Texture2D((int)width, (int)height);
 
@@ -45,15 +66,19 @@ namespace NRPFarmod.ContentManager {
 
                     index++;
                 } while (enumerator.MoveNext());
-
                 tmp.Apply();
-                MelonLogger.Msg($"Texture: \u001b[32mok\u001b[0m");
-            } catch(Exception) {
+                //MelonLogger.Msg($"Texture: \u001b[32mok\u001b[0m");
+            } catch (Exception) {
                 MelonLogger.Msg($"Texture: \u001b[31mERROR\u001b[0m");
             }
-           
+
         }
 
+        /// <summary>
+        /// Ließt die Color Werte aus dem Bytearray aus
+        /// </summary>
+        /// <param name="obj"></param>
+        /// <returns></returns>
         public static IEnumerator<uint> LoadValuesFromByte(byte[] obj) {
             using (var stream = new MemoryStream(obj))
             using (var reader = new BinaryReader(stream)) {
@@ -64,7 +89,11 @@ namespace NRPFarmod.ContentManager {
             }
         }
 
-
+        /// <summary>
+        /// Bitmanipulation
+        /// </summary>
+        /// <param name="blob"></param>
+        /// <returns></returns>
         public static (byte, byte, byte, byte) ShiftOut(uint blob) {
             byte a = (byte)(blob & 0xFF);
             byte b = (byte)((blob >> 8) & 0xFF);
@@ -72,5 +101,6 @@ namespace NRPFarmod.ContentManager {
             byte d = (byte)((blob >> 24) & 0xFF);
             return (a, b, c, d);
         }
+        #endregion
     }
 }
